A new package is added here jdeqsimPerformance.
A new LogEnterLinkEvents.java class is added that will log the event whenever a link will be entered during our simulation run. At the moment this class is just a stub.

The Main_equil_bareBone.java class is just copied from jdeqsim package and modified to add the new handler that is LogEnterLinkEvents.java to it.