Implement one-way link logic in writing a MATSim network
The R5 network has two edges for every node pair, regardless of the topology of the edge. This implements a logical check where the forward direction is always added to the MATSim network, but the backward direction is not added for one-way links.

There is a remaining TODO regarding links whose directionality is reverse from their topology.

Addresses #220