Modifies hasValidChargingCapability filter.
Previously, the filter would treat EVs and petrol vehicles the same way. And would not care about the activity type either (petrol vehicles were fine sine this would return true).

Also, the filter would check if the charger can provide a power smaller than or equal to what the vehicle is rated for, which is the opposite desired effect. We want the charger to be able to provide the vehicle's rated charging power or more.

The changes made verify if this is an EV which is in an activity that requires charging (charge or enroute), and then verifies if the charger in question has the rated power required, otherwise it always returns true.