Adds more coverage to the tests.
More changing stations were added that should not be suitable for the BEVs to charge (not enough power rating).

Adds a randomized seed to the test instead of being a constant one. Had to crank up prices to 200 USD on expensive charging stations to consistently have them used less than the free ones (might need to discuss parameter tuning).