Changes "EmbodiedBeamLeg.splitLegForParking" (source of another cost discrepancy) and "currentTripCosts" from "BasePersonData":
Fixes "EmbodiedBeamLeg.splitLegForParking" which now properly evaluates the cost of each part of the splitted legs instead of assigning their full cost to the first part of the leg.

Removes "currentTripCosts: Double" field from BasePersonData and moves it to "def currentTripCosts()" that obtains the cost from "currentTrip: Option[EmbodiedBeamTrip]".

This field was causing unnecessary complexity to BasePersonData, since if at any point of the code currentTrip was modified but currentTripCosts was not updated to reflect these changes, these values would get out of sync.

Still have to increase PersonAgentTripCostTest coverage to see if the changes made on this commit that involve en-route code on PersonAgent.scala#L972 do not have adverse effects.